moved platform_guard.h to uil; defined UT_PLATFORM() macro
