+++ src/uTensor/ops/optimized/x86/FFT.hpp
@@ -1,8 +1,8 @@
 #ifndef UTENSOR_FFT_HPP
 #define UTENSOR_FFT_HPP
 
-#include "uTensor/core/platform_guard.hpp"
-#if UTENSOR_PLATFORM == UTENSOR_PLATFORM_X86
+#include "uTensor/util/platform_guard.hpp"
+#if UT_PLATFORM(UT_PLATFORM_X86)
 
 #include "fftw3.h"
 #include "operatorBase.hpp"
